add tests

+++ src/main/java/dev/zarr/zarrjava/utils/IndexingUtils.java
@@ -87,7 +87,7 @@ public class IndexingUtils {
 
             if (selOffset[dimIdx] + selShape[dimIdx] > dimLimit) {
                 // selection ends after current chunk
-                shape[dimIdx] = (int) (chunkShape[dimIdx] - chunkOffset[dimIdx]);
+                shape[dimIdx] = chunkShape[dimIdx] - chunkOffset[dimIdx];
             } else {
                 // selection ends within current chunk
                 shape[dimIdx] = (int) (selOffset[dimIdx] + selShape[dimIdx] - dimOffset
